LayoutTests/fast/forms/ios/select-picker-change-and-focus-another-select-expected.txt

This test verifies changing the value of one select element using the picker and then focusing another select element only changes the value of the first element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS first.value is "B"
PASS second.value is "1"
PASS successfullyParsed is true

TEST COMPLETE
  
